Safari-609.1.11.1/LayoutTests/http/tests/storageAccess/request-and-grant-access-then-detach-should-not-have-access-expected.txt

Tests that a cross-origin iframe from a prevalent domain that is granted storage access and then is detached from the DOM does not have storage access.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Storage access was granted. document.cookie == firstPartyCookie=value, cookies seen server-side == {"firstPartyCookie":"value"}
PASS There is a storage access entry for localhost.
PASS There is no storage access entry for localhost after iframe detach.
PASS successfullyParsed is true

TEST COMPLETE
